feat(contract): modify how we handle depdencies
I believe this is a better way to manage dependencies accross the contract. The main.lua will set global functions in state, but they can reference or require them directly if they are not available. This makes testing trivial and aligns better with how you would write the contract using classes and normal DI.