fix(entropy): fix observer entropy by applying sorting on observers before prescribing
This is a bug in our smartweave implementation that Lua caught. Lua does not guranatee the order of processing an array, so we must sort them using entropy provided by the hashchain combined with the gateway addresses. This will make sure that the order in which observers are prescribed are deterministic for an epoch.